refactor: Extract file engine Extract translate engine Extract interface
